+++ resources/js/views/AdminSettings.vue
@@ -140,7 +140,9 @@
                   $t("admin.settings.pagination_page_size.description")
                 }}</small>
                 <FormError
-                  :errors="formErrors.fieldError('pagination_page_size')"
+                  :errors="
+                    formErrors.fieldError('general_pagination_page_size')
+                  "
                 />
               </div>
             </div>
@@ -1492,7 +1494,9 @@ function updateSettings() {
 
   if (uploadBBBLogoFile.value) {
     formData.append("bbb_logo_file", uploadBBBLogoFile.value);
-  } else if (!bbbLogoDeleted.value && settings.value.bbb_logo != null) {
+  } else if (bbbLogoDeleted.value) {
+    formData.append("bbb_logo", "");
+  } else if (settings.value.bbb_logo !== null) {
     formData.append("bbb_logo", settings.value.bbb_logo);
   }